# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/usuarios/registro/cliente": {
      "post": {
        "operationId": "UsuariosController_crearCliente",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearClienteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios/registro/vendedor": {
      "post": {
        "operationId": "UsuariosController_crearVendedor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearVendedorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios/registro/demovendedor": {
      "post": {
        "operationId": "UsuariosController_crearDemoVendedor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearDemoVendedorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Para crear usuario de DEMOVENDEDOR, se usa en formulario de UNETE A NOSOTROS",
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios/registro/trabajador": {
      "post": {
        "operationId": "UsuariosController_crearTrabajador",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearTrabajadorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios/actualizar/info": {
      "patch": {
        "operationId": "UsuariosController_actualizarInfo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualiza toda la info de un cliente",
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios/actualizar/direccion": {
      "patch": {
        "operationId": "UsuariosController_actualizarDireccion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarDireccionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualiza especificamente la direccion de un usuario",
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios/actualizar/foto": {
      "patch": {
        "operationId": "UsuariosController_actualizarFotoPerfil",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarFotoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualiza especificamente la foto de un usuario",
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios/registro/admin": {
      "post": {
        "operationId": "UsuariosController_crearAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPrimerAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Crea un nuevo usuario ADMIN",
        "tags": [
          "Usuarios"
        ]
      }
    },
...
  },
  "info": {
    "title": "MarketJoyas",
    "description": "Documentacion de endpoints de marketjoyas",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CrearClienteDto": {
        "type": "object",
        "properties": {
          "nombre_completo": {
            "type": "string",
            "description": "Nombre completo del cliente",
            "example": "María García López"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del cliente",
            "example": "maria.garcia@email.com"
          },
          "direccion": {
            "type": "string",
            "description": "Dirección del cliente (opcional)",
            "example": "Av. Principal 123, Lima, Perú (opcional)"
          },
          "dni": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8,
            "description": "DNI del cliente (8 caracteres)",
            "example": "71234567"
          },
          "telefono": {
            "type": "string",
            "description": "Número de teléfono del cliente",
            "example": "+51987654321"
          },
          "contraseña": {
            "type": "string",
            "description": "Contraseña del cliente",
            "example": "ClaveSegura123",
            "minLength": 6
          }
        },
        "required": [
          "nombre_completo",
          "email",
          "dni",
          "telefono",
          "contraseña"
        ]
      },
      "TiendaDto": {
        "type": "object",
        "properties": {
          "nombre_tienda": {
            "type": "string",
            "description": "Nombre de la tienda",
            "example": "Tienda Central"
          },
          "direccion_tienda": {
            "type": "string",
            "description": "Dirección de la tienda",
            "example": "Av. Comercio 456"
          },
          "telefono_tienda": {
            "type": "string",
            "description": "Teléfono de la tienda",
            "example": "+5112345678"
          },
          "pais": {
            "type": "string",
            "description": "País donde se encuentra la tienda",
            "example": "Perú"
          },
          "ciudad": {
            "type": "string",
            "description": "Ciudad donde se encuentra la tienda",
            "example": "Lima"
          },
          "provincia": {
            "type": "string",
            "description": "Provincia donde se encuentra la tienda",
            "example": "Lima"
          },
          "codigoPostal": {
            "type": "string",
            "description": "Código postal de la tienda",
            "example": "15001"
          }
        },
        "required": [
          "nombre_tienda"
        ]
      },
      "CrearVendedorDto": {
        "type": "object",
        "properties": {
          "nombre_completo": {
            "type": "string",
            "description": "Nombre completo del vendedor",
            "example": "Juan Pérez Gómez"
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario para el vendedor",
            "example": "juanperez123"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del vendedor",
            "example": "juan.perez@email.com"
          },
          "direccion": {
            "type": "string",
            "description": "Dirección del vendedor",
            "example": "Calle Falsa 123, Ciudad"
          },
          "dni": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8,
            "description": "DNI del vendedor (8 caracteres)",
            "example": "12345678"
          },
          "telefono": {
            "type": "string",
            "description": "Número de teléfono del vendedor",
            "example": "+51987654321"
          },
          "contraseña": {
            "type": "string",
            "description": "Contraseña del vendedor",
            "example": "admin123",
            "minLength": 6
          },
          "tienda": {
            "description": "Datos de la tienda asociada",
            "allOf": [
              {
                "$ref": "#/components/schemas/TiendaDto"
              }
            ]
          }
        },
        "required": [
          "email",
          "contraseña"
        ]
      },
      "CrearDemoVendedorDto": {
        "type": "object",
        "properties": {
          "nombre_completo": {
            "type": "string",
            "description": "Nombre completo del usuario",
            "example": "Juan Pérez García"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del usuario",
            "example": "juanperez@email.com"
          },
          "telefono": {
            "type": "string",
            "description": "Número de teléfono",
            "example": "+51987654321"
          },
          "direccion": {
            "type": "string",
            "description": "Dirección del usuario (opcional)",
            "example": "Av. Siempre Viva 742"
          },
          "nombre_tienda": {
            "type": "string",
            "description": "Nombre de la tienda",
            "example": "Repuestos El Rápido"
          },
          "direccion_tienda": {
            "type": "string",
            "description": "Dirección fiscal de la tienda (opcional)",
            "example": "Jr. Comercio 123, Cercado de Lima"
          },
          "departamento": {
            "type": "string",
            "description": "Departamento donde se ubica la tienda (opcional)",
            "example": "Lima"
          },
          "provincia": {
            "type": "string",
            "description": "Provincia donde se ubica la tienda (opcional)",
            "example": "Lima"
          },
          "distrito": {
            "type": "string",
            "description": "Distrito donde se ubica la tienda (opcional)",
            "example": "San Isidro"
          },
          "contraseña": {
            "type": "string",
            "minLength": 6,
            "description": "Contraseña del usuario (mínimo 6 caracteres)",
            "example": "miSuperClave123"
          }
        },
        "required": [
          "nombre_completo",
          "email",
          "telefono",
          "nombre_tienda",
          "contraseña"
        ]
      },
      "CrearTrabajadorDto": {
        "type": "object",
        "properties": {
          "nombre_completo": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "contraseña": {
            "type": "string",
            "minLength": 6
          },
          "telefono": {
            "type": "string"
          },
          "dni": {
            "type": "string"
          }
        },
        "required": [
          "nombre_completo",
          "email",
          "contraseña",
          "telefono",
          "dni"
        ]
      },
      "ActualizarInfoDto": {
        "type": "object",
        "properties": {
          "nombre_completo": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "dni": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8,
            "example": "12345678"
          },
          "telefono": {
            "type": "string",
            "example": "+51987654321"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "juan@email.com"
          },
          "contraseña_actual": {
            "type": "string",
            "example": "MiClaveActual123"
          },
          "nueva_contraseña": {
            "type": "string",
            "example": "MiNuevaClave456"
          }
        }
      },
      "ActualizarDireccionDto": {
        "type": "object",
        "properties": {
          "direccion": {
            "type": "string",
            "example": "Av. Siempre Viva 742"
          }
        }
      },
      "ActualizarFotoDto": {
        "type": "object",
        "properties": {
          "fotoPerfilUrl": {
            "type": "string",
            "example": "https://res.cloudinary.com/miimagen.jpg"
          },
          "fotoPerfilPublicId": {
            "type": "string",
            "example": "cloudinary-public-id-123 o lo que devuelva"
          }
        },
        "required": [
          "fotoPerfilUrl",
          "fotoPerfilPublicId"
        ]
      },
      "CrearPrimerAdminDto": {
        "type": "object",
        "properties": {
          "nombre_completo": {
            "type": "string",
            "description": "Nombre completo del administrador",
            "example": "Juan Pérez Torres"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del administrador",
            "example": "juan.perez@admin.com"
          },
          "direccion": {
            "type": "string",
            "description": "Dirección del administrador (opcional)",
            "example": "Calle 45 #123, Bogotá, Colombia"
          },
          "dni": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8,
            "description": "DNI del administrador (8 caracteres)",
            "example": "87654321"
          },
          "telefono": {
            "type": "string",
            "description": "Número de teléfono del administrador",
            "example": "+573001112233"
          },
          "contraseña": {
            "type": "string",
            "description": "Contraseña del administrador (mínimo 6 caracteres)",
            "example": "AdminFuerte2024",
            "minLength": 6
          }
        },
        "required": [
          "nombre_completo",
          "email",
          "dni",
          "telefono",
          "contraseña"
        ]
      },
...
    }
  }
}